pythonlibs/mantis/requirements.txt

kafka-python
flask
#celery
pykafka
flask-login
flask-openid
flask-mail
flask-sqlalchemy
sqlalchemy-migrate
#flask-whooshalchemy
flask-wtf
#flask-babel
#virtualenv
gevent
psycopg2
#uwsgi
#flask-cors
#pyDes
#jpype1
requests
psycogreen
#gevent-psycopg2
redis
xpinyin
#python-qpid
PyYAML
python-dateutil
Flask-CAS
#kazoo
Flask-Compress
Flask-Gzip
#Flask-MySQLDb
flask-cors
psutil
pymongo
websocket-client
msgpack-python
#qdarkstyle
#SortedContainers
#futuquant
wmi
#future
#flask-socketio
flask-restful
flask-cors
gevent-websocket
#ta-lib
grequests
pp  # https://www.parallelpython.com/content/view/15/30/#QUICKSMP
redlock-py
fire
flask_sqlalchemy
flask_compress
gevent-socketio

#qpid-python-1.36.0.tar.gz